rbd: replace deprecated rbd_get_parent_info() with rbd_get_parent()
